requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.0.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.37.0
coloredlogs==15.0.1
cookiecutter==2.5.0
crispy-bootstrap5==0.7
crispy-tailwind==0.5.0
cryptography==39.0.1
cssselect==1.2.0
dataclasses-json==0.6.4
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==4.2
django-allauth==0.52.0
django-appconf==1.0.6
django-browser-reload==1.7.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.11.2
django-htmx==1.17.2
django-tagulous==1.3.3
django-tailwind==3.5.0
django-tinymce==3.7.1
django-vectordb==0.2.0
django-widget-tweaks==1.5.0
fastapi==0.110.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.1.2
install==1.3.5
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.4
kubernetes==29.0.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langsmith==0.1.8
llama-index==0.10.16
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.7
llama-index-core==0.10.16.post1
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.7
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.5
llama-index-vector-stores-pinecone==0.1.4
llama-parse==0.3.6
llamaindex-py-client==0.1.13
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
pinecone-client==3.1.0
posthog==3.5.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psycopg==3.1.8
psycopg-binary==3.1.8
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.6.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.3
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==12.0
whitenoise==6.4.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
